pyproject.toml
[project]
name = "atrium-csd-ner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argilla>=2.8.0",
    "datasets>=3.6.0",
    "dotenv>=0.9.9",
    "evaluate>=0.4.4",
    "fastapi>=0.116.1",
    "flair>=0.15.1",
    "gliner2>=1.0.2",
    "html2text>=2025.4.15",
    "huggingface-hub>=0.32.4",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "itables>=2.5.2",
    "joblib>=1.5.1",
    "jupyter>=1.1.1",
    "langchain-core>=0.3.66",
    "langchain-openai>=0.3.25",
    "litellm>=1.72.6.post1",
    "multidict>=6.5.0",
    "nltk>=3.9.1",
    "openai>=1.88.0",
    "pandas>=2.3.0",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "seqeval>=1.2.2",
    "spacy>=3.8.7",
    "streamlit>=1.46.0",
    "streamlit-autorefresh>=1.0.1",
    "tabulate>=0.9.0",
    "transformers>=4.52.4",
    "uvicorn>=0.34.3",
    "wordcloud>=1.9.4",
    "wtpsplit>=1.3.0",
]

